chore: introduce more active maintainers
I've done this in the reverse and made an "executive decision" to add these co-maintainers. 

Usually the process would be to propose in a PR as such, and then once approved, add them, but I wanted folks to be able to hit the ground running here.

Moving forward, I would suggest giving new maintainer applicants "triage" privilege to start, and then promoting them to "write" once they've gained enough experience helping community members, assisting with building the roadmap, and helping to review or introduce PRs.